examples/01_code_generation/requirements.txt
pytest>=7.4.0
pytest-cov>=4.1.0
dnspython>=2.4.0
